Revert "Use SerializableTransactionManager instead of Snapshot"
This isn't a problem by itself, but I'm no longer going to go with using
serializable transactions as I previously wanted to. Reverting as this affects
some upstream code.

This reverts commits:
e43392c2b8f09555d38c1eefdc8a7d8d018dc89d
9849ad4467726f9a6788f033602bca7c1976cc9f
2c7ad022dc030a530f5cfde4a278bdbf77ecf59f
